Frogger: Move FroggerGame class into a separate file in the 'game' directory.
